Migrate helper methods
