+++ Mappalachia/Class/Helpers/DataHelper.cs
@@ -5,7 +5,7 @@ using System.Linq;
 namespace Mappalachia
 {
 	// Provides data helper methods, data translation and sorting
-	static class DataHelper
+	public static class DataHelper
 	{
 		// A list of entities which are often (mis)represented instead by LVLI in the data
 		public static readonly List<string> typicalLVLIItems = new List<string> { "FLOR", "ALCH", "WEAP", "ARMO", "BOOK", "AMMO" };
@@ -280,6 +280,24 @@ namespace Mappalachia
 		public static double GetSpawnChance(string signature, string editorID)
 		{
 			return (signature == "LVLI" || editorID.Contains("ChanceNone")) ? -1 : 100;
+		}
+
+		// returns the nth item in a list as if it were cyclic (supports <0 or >n)
+		public static T GetCyclicItem<T>(List<T> collection, int n)
+		{
+			if (collection.Count == 0)
+			{
+				return default;
+			}
+
+			n %= collection.Count;
+
+			if (n < 0)
+			{
+				n = collection.Count + n;
+			}
+
+			return collection[n];
 		}
 	}
 }
