+++ Mappalachia/Class/Mappalachia.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Diagnostics;
 using System.Windows.Forms;
 using Mappalachia.Class;
 
@@ -26,11 +25,6 @@ namespace Mappalachia
 
 				// Choosing to not save settings, since we just crashed it's likely safer to not keep them.
 			}
-		}
-
-		public static void LaunchURL(string url)
-		{
-			Process.Start(new ProcessStartInfo { FileName = url, UseShellExecute = true });
 		}
 	}
 }
